create board- and gameWrapper
